# openapi.yaml
openapi: 3.0.3
info:
  title: Vokter API
  version: 1.0.0
  description: API de comercio electrónico para clientes web y móvil.
servers:
  - url: http://localhost:5000/api/v1
...
paths:
...
  /catalog/products:
    get:
      tags: [Catalog]
      summary: List active products
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: featured, in: query, schema: {type: boolean, default: false}}
        - {name: category, in: query, schema: {type: string}}
        - {name: brand, in: query, schema: {type: string}}
        - {name: min_price, in: query, schema: {type: number, format: double, minimum: 0}}
        - {name: max_price, in: query, schema: {type: number, format: double, minimum: 0}}
        - {name: available, in: query, schema: {type: boolean, default: false}}
        - {name: sort, in: query, schema: {type: string, enum: [newest, price_asc, price_desc, name], default: newest}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Product'}
                  meta: {$ref: '#/components/schemas/PaginationMeta'}
  /catalog/products/{product_id}:
    get:
      tags: [Catalog]
      summary: Get a product with variants and images
      parameters:
        - {$ref: '#/components/parameters/ProductId'}
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/ProductDetail'}
        '404': {$ref: '#/components/responses/NotFoundError'}
  /catalog/products/slug/{slug}:
    get:
      tags: [Catalog]
      summary: Get a product by slug
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/ProductDetail'}
        '404': {$ref: '#/components/responses/NotFoundError'}
...
  /carts:
    post:
      tags: [Cart]
      summary: Create or retrieve a cart
      description: Authenticated requests use the current user. Anonymous requests use X-Cart-Session or session_key.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartSessionHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCartRequest'
      responses:
        '201':
          description: Cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /carts/current:
    get:
      tags: [Cart]
      summary: Get the authenticated user's active cart
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Current cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /carts/merge:
    post:
      tags: [Cart]
      summary: Merge an anonymous cart into the authenticated cart
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeCartRequest'
      responses:
        '200':
          description: Merged cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/ConflictError'}
  /carts/{cart_id}:
    get:
      tags: [Cart]
      summary: Get a cart
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartId'}
        - {$ref: '#/components/parameters/CartSessionHeader'}
      responses:
        '200':
          description: Cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '404': {$ref: '#/components/responses/NotFoundError'}
  /carts/{cart_id}/items:
    post:
      tags: [Cart]
      summary: Add a variant to a cart
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartId'}
        - {$ref: '#/components/parameters/CartSessionHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '201':
          description: Updated cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/ConflictError'}
    delete:
      tags: [Cart]
      summary: Remove all items from a cart
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartId'}
        - {$ref: '#/components/parameters/CartSessionHeader'}
      responses:
        '200':
          description: Empty cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
  /carts/{cart_id}/items/{item_id}:
    patch:
      tags: [Cart]
      summary: Update cart item quantity
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartId'}
        - {$ref: '#/components/parameters/ItemId'}
        - {$ref: '#/components/parameters/CartSessionHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/ConflictError'}
    delete:
      tags: [Cart]
      summary: Remove a cart item
      security:
        - bearerAuth: []
        - {}
      parameters:
        - {$ref: '#/components/parameters/CartId'}
        - {$ref: '#/components/parameters/ItemId'}
        - {$ref: '#/components/parameters/CartSessionHeader'}
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Cart'}
...
  /admin/catalog/categories:
    post:
      tags: [Administration]
      summary: Create a category
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryRequest'}
      responses:
        '201': {$ref: '#/components/responses/CategorySuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /admin/catalog/categories/{category_id}:
    patch:
      tags: [Administration]
      summary: Update a category
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/CategoryId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CategoryUpdateRequest'}
      responses:
        '200': {$ref: '#/components/responses/CategorySuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
    delete:
      tags: [Administration]
      summary: Deactivate a category
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/CategoryId'}
      responses:
        '200': {$ref: '#/components/responses/CategorySuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /admin/catalog/products:
    post:
      tags: [Administration]
      summary: Create a product
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProductRequest'}
      responses:
        '201': {$ref: '#/components/responses/ProductDetailSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /admin/catalog/products/{product_id}:
    patch:
      tags: [Administration]
      summary: Update a product
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProductId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProductUpdateRequest'}
      responses:
        '200': {$ref: '#/components/responses/ProductDetailSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
    delete:
      tags: [Administration]
      summary: Deactivate a product and its variants
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProductId'}
      responses:
        '200': {$ref: '#/components/responses/ProductDetailSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /admin/catalog/products/{product_id}/variants:
    post:
      tags: [Administration]
      summary: Create a product variant
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProductId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VariantRequest'}
      responses:
        '201': {$ref: '#/components/responses/VariantSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /admin/catalog/variants/{variant_id}:
    patch:
      tags: [Administration]
      summary: Update a variant and inventory
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/VariantId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VariantUpdateRequest'}
      responses:
        '200': {$ref: '#/components/responses/VariantSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      name: product_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    CategoryId:
      name: category_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    VariantId:
      name: variant_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    CartId:
      name: cart_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
    ItemId:
      name: item_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    CartSessionHeader:
      name: X-Cart-Session
      in: header
      required: false
      schema: {type: string, maxLength: 100}
...
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, is_active]
      properties:
        id: {type: integer}
        name: {type: string}
        slug: {type: string}
        description: {type: string, nullable: true}
        is_active: {type: boolean}
    Product:
      type: object
      required: [id, name, slug, base_price, is_active, is_featured, categories]
      properties:
        id: {type: integer}
        name: {type: string}
        slug: {type: string}
        description: {type: string, nullable: true}
        brand: {type: string, nullable: true}
        base_price: {type: string, pattern: '^\d+\.\d{2}$'}
        is_active: {type: boolean}
        is_featured: {type: boolean}
        categories:
          type: array
          items: {$ref: '#/components/schemas/Category'}
    ProductDetail:
      allOf:
        - {$ref: '#/components/schemas/Product'}
        - type: object
          required: [images, variants]
          properties:
            images:
              type: array
              items: {$ref: '#/components/schemas/ProductImage'}
            variants:
              type: array
              items: {$ref: '#/components/schemas/ProductVariant'}
    ProductRequest:
      type: object
      required: [name, slug, base_price]
      properties:
        name: {type: string, maxLength: 180}
        slug: {type: string, maxLength: 200}
        description: {type: string, nullable: true}
        brand: {type: string, nullable: true}
        base_price: {type: string, pattern: '^\d+(\.\d{1,2})?$'}
        is_active: {type: boolean, default: true}
        is_featured: {type: boolean, default: false}
        category_ids:
          type: array
          items: {type: integer, minimum: 1}
    ProductUpdateRequest:
      type: object
      properties:
        name: {type: string, maxLength: 180}
        slug: {type: string, maxLength: 200}
        description: {type: string, nullable: true}
        brand: {type: string, nullable: true}
        base_price: {type: string, pattern: '^\d+(\.\d{1,2})?$'}
        is_active: {type: boolean}
        is_featured: {type: boolean}
        category_ids:
          type: array
          items: {type: integer, minimum: 1}
    CategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name: {type: string, maxLength: 100}
        slug: {type: string, maxLength: 120}
        description: {type: string, nullable: true}
        is_active: {type: boolean, default: true}
    CategoryUpdateRequest:
      type: object
      properties:
        name: {type: string, maxLength: 100}
        slug: {type: string, maxLength: 120}
        description: {type: string, nullable: true}
        is_active: {type: boolean}
    VariantRequest:
      type: object
      required: [sku, name, price, stock_quantity]
      properties:
        sku: {type: string, maxLength: 80}
        name: {type: string, maxLength: 120}
        price: {type: string, pattern: '^\d+(\.\d{1,2})?$'}
        stock_quantity: {type: integer, minimum: 0}
        is_active: {type: boolean, default: true}
    VariantUpdateRequest:
      type: object
      properties:
        sku: {type: string, maxLength: 80}
        name: {type: string, maxLength: 120}
        price: {type: string, pattern: '^\d+(\.\d{1,2})?$'}
        stock_quantity: {type: integer, minimum: 0}
        is_active: {type: boolean}
    ProductImage:
      type: object
      required: [id, url, sort_order]
      properties:
        id: {type: integer}
        url: {type: string, format: uri}
        alt_text: {type: string, nullable: true}
        sort_order: {type: integer}
    ProductVariant:
      type: object
      required: [id, sku, name, price, stock_quantity, is_active]
      properties:
        id: {type: integer}
        sku: {type: string}
        name: {type: string}
        price: {type: string, pattern: '^\d+\.\d{2}$'}
        stock_quantity: {type: integer, minimum: 0}
        is_active: {type: boolean}
    Cart:
      type: object
      required: [id, user_id, session_key, status, items, total]
      properties:
        id: {type: integer}
        user_id: {type: integer, nullable: true}
        session_key: {type: string, nullable: true}
        status: {type: string, enum: [active, checked_out]}
        items:
          type: array
          items: {$ref: '#/components/schemas/CartItem'}
        total: {type: string, pattern: '^\d+\.\d{2}$'}
    CartItem:
      type: object
      required: [id, variant_id, quantity, unit_price]
      properties:
        id: {type: integer}
        variant_id: {type: integer}
        quantity: {type: integer, minimum: 1}
        unit_price: {type: string, pattern: '^\d+\.\d{2}$'}
    CreateCartRequest:
      type: object
      properties:
        session_key: {type: string, maxLength: 100}
    MergeCartRequest:
      type: object
      required: [session_key]
      properties:
        session_key: {type: string, minLength: 1, maxLength: 100}
    AddCartItemRequest:
      type: object
      required: [variant_id]
      properties:
        variant_id: {type: integer, minimum: 1}
        quantity: {type: integer, minimum: 1, default: 1}
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity: {type: integer, minimum: 1}
...
    PaginationMeta:
      type: object
      required: [page, per_page, count, total, pages]
      properties:
        page: {type: integer}
        per_page: {type: integer}
        count: {type: integer}
        total: {type: integer}
        pages: {type: integer}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string, format: uuid, nullable: true}
  responses:
...
    CategorySuccess:
      description: Category response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {$ref: '#/components/schemas/Category'}
    ProductDetailSuccess:
      description: Product detail response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {$ref: '#/components/schemas/ProductDetail'}
    VariantSuccess:
      description: Product variant response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {$ref: '#/components/schemas/ProductVariant'}
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ConflictError:
      description: Business or data conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}